Fix issue #143 (#144)
We were optimizing when the height of the layout increased more than the actual layout height already computed, by skipping a remeasure. But it turns out that we don't correctly flag the case where the original measure was too small. Disabling the optimization for now as we need to keep track of "too small" not only for the container but also for individual widgets that have been measured.

- added a OPTIMIZE_HEIGHT_CHANGE flag
- added check 420