Make ConstraintLayout an inline composable
This is consistent with how most layouts in foundation-layout are
implemented. In addition, it implicitly fixes issue 181717954, which was
happening because the content lambda could previously be recomposed
without recomposing the parent content lambda given to
MultiMeasureLayout - therefore, the scope was not reset() but it was
added the barrier again causing unexpected behavior.

Fixes: 181717954
Test: ConstraintLayoutTest