Make ConstraintSet objects extendable
It is now possible to extend a new ConstraintSet written in DSL using a
ConstraintSet written in Json and vice-versa.

This is so that ConstraintSet themselves are more interchangeable and
the developer can write their new ConstraintSet in their preferred
syntax.

Modified visibility for related classes that should be internal.

Reorganized some classes in different files (package is umodified,
should not affect built classes).