[ComposeMotion] Fix custom properties
Custom properties would return values that didn't correspond to the
layout, since they were effectively being calculated with a progress
value form a different measure pass.

This makes it so that custom properties are always calculated with the
current progress value used to measure the layout in the current
composition.