Updated Material2DAnimationSpec
It's now possible to provide a second easing curve to handle
interpolations for the 3rd and 4th variable.

This is usefuld for example when interpolating bounds, where the first
two variables may correpond to the location and the last two variables
correspond to the dimensions and you may want different easing curves
for each pair.